Split and naming conventions
Split the file into two parts and used naming conventions (s1 stage1, s2 stage2)
